Se implementa nuevos ajustes a Recaudo y Filtro de Report de despacho por rango de fehcas

+++ database/migrations/2020_07_29_003328_create_route_takings_table.php
@@ -17,7 +17,7 @@ class CreateRouteTakingsTable extends Migration
             $table->increments('id');
             $table->integer('total_production')->nullable();
             $table->integer('control')->nullable();
-            $table->integer('fuel')->nullable();
+            $table->double('fuel')->nullable();
             $table->integer('others')->nullable();
             $table->integer('net_production')->nullable();
             $table->text('observations')->nullable();
